+++ R/pattern-geometry-tiling.R
@@ -24,6 +24,7 @@
 #' \item{truncated_hexagonal}{Creates a truncated hexagonal tiling made of dodecagons and triangles.}
 #' \item{truncated_trihexagonal}{Creates a truncated trihexagonal tiling made of hexagons, squares, and triangles.}
 #' \item{`3.3.8*15.3.4.3.8*15`}{Creates a regular (star) polygon tiling made of triangles, squares, and eight-pointed stars.}
+#' \item{`3.4.8.3.8*15`}{Creates a regular (star) polygon tiling made of triangles, squares, octagons, and eight-pointed stars.}
 #' }
 #'
 #' @inheritParams grid.pattern_circle
@@ -93,19 +94,8 @@ names_polygon_tiling <- c("herringbone",
                           "truncated_square",
                           "truncated_hexagonal",
                           "truncated_trihexagonal",
-                          "3.3.8*15.3.4.3.8*15")
-
-# 3.3.8*15.3.4.3.8*15
-
-# More tilings we could do
-# What should their names be
-# Hexagonal replaced with star?
-# Truncated Square with (alternating) star?
-# Truncated Hexagonal with star?
-# Trihexagonal topological equivalent "star6"
-# Trihexagonal topological equivalent "star3"
-# Trihexagonal topological equivalent "convex3"
-# Lots of stars...
+                          "3.3.8*15.3.4.3.8*15",
+                          "3.4.8.3.8*15")
 
 create_pattern_polygon_tiling <- function(params, boundary_df, aspect_ratio, legend = FALSE) {
     type <- tolower(params$pattern_type)
@@ -138,6 +128,7 @@ create_pattern_polygon_tiling <- function(params, boundary_df, aspect_ratio, leg
                  truncated_square = create_trunc_square_tiling(xyi, gp, spacing, angle),
                  truncated_trihexagonal = create_trunc_trihex_tiling(xyi, gp, spacing, angle),
                  `3.3.8*15.3.4.3.8*15` = create_3.3.8_15.3.4.3.8_15_tiling(xyi, gp, spacing, angle),
+                 `3.4.8.3.8*15` = create_3.4.8.3.8_15_tiling(xyi, gp, spacing, angle),
                  abort(paste("Don't know how to do tiling", type)))
     gTree(children = gl, name = "polygon_tiling")
 }
@@ -166,7 +157,31 @@ create_3.3.8_15.3.4.3.8_15_tiling <- function(xyi, gp, spacing, angle) {
                         spacing = spacing, angle = angle, gp = gp_star,
                         name = "stars")
     gList(bg, octagons, stars)
+}
 
+create_3.4.8.3.8_15_tiling <- function(xyi, gp, spacing, angle) {
+    n_col <- length(gp$fill)
+    gp_star <- gp_oct <- gp_bg <- gp
+    if (n_col == 2L) {
+        gp_bg$fill <- gp_star$fill <- gp$fill[2L]
+        gp_oct$fill <- gp$fill[1L]
+    } else if (n_col == 3L) {
+        gp_bg$fill <- gp$fill[2]
+        gp_oct$fill <- gp$fill[c(1,3)]
+        gp_star$fill <- gp$fill[2]
+    }
+    bg <- polygonGrob(xyi$x, xyi$y, xyi$id, default.units = "npc", gp = gp_bg,
+                      name = "background_color")
+    octagons <- patternGrob("regular_polygon", xyi$x, xyi$y, xyi$id,
+                            shape = "convex8", density = 1.082, rot = 22.5,
+                            spacing = spacing, angle = angle, gp = gp_oct,
+                            name = "octagons")
+    scale <- star_scale(8, 60, external = TRUE)
+    stars <- patternGrob("regular_polygon", xyi$x, xyi$y, xyi$id,
+                         shape = c("null", "star8"), density = 1.082, rot = 22.5,
+                        spacing = spacing, angle = angle, gp = gp_star,
+                        name = "stars")
+    gList(bg, octagons, stars)
 }
 
 create_herringbone_tiling <- function(xyi, gp, spacing, angle) {